move type utils to dedicated private module
